position-details

+++ src/main/java/com/springboot/services/PositionService.java
@@ -19,18 +19,52 @@ public class PositionService implements IPositionService {
 	}
 
 	@Override
-	public Position addPosition(Position p) {
-		return positionRepository.save(p);
-	}
-
-	@Override
-	public void deletePosition(Integer id) {
-		positionRepository.deleteById(id);
-	}
-
-	@Override
 	public Position retrievePosition(Integer id) {
 		return positionRepository.findById(id).orElse(null);
+	}
+	
+	@Override
+	public String addPosition(Position p) {
+		try {
+			positionRepository.save(p);
+			return "Cette position a été sauvegardée";
+		} catch (Exception e) {
+			throw e;
+		}
+	}
+
+	@Override
+	public String updatPosition(Position p) {
+		try {
+			Position position = positionRepository.findById(p.getId()).orElse(null);
+			if(position != null) {
+				positionRepository.save(position);
+				return "Cette position a été modifiée";
+			}else {
+				return "Cette position n'existe pas";
+			}
+		} catch (Exception e) {
+			throw e;
+		}
+	}
+
+	@Override
+	public String deletPosition(Integer id) {
+		try {
+			if(positionRepository.findById(id).orElse(null) != null) {
+				positionRepository.deleteById(id);
+				return "Cette position a été supprimé";
+			}else {
+				return "Cette position n'existe pas";
+			}
+		} catch (Exception e) {
+			throw e;
+		}
+	}
+
+	@Override
+	public Position retrievePositionByLatAndLong(String latitude, String longitude) {
+		return positionRepository.findByLatitudeAndLongitude(latitude, longitude);
 	}
 
 }
